site/pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "VLLM-enabled library for extracting images and tables from scientific documents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "huggingface>=0.0.1",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "transformers>=4.49.0",
]
